requirements.txt
setuptools
albumentations
numpy
opencv-python  # Required by albumentations
torch
torchvision
Pillow
pycocotools